project/requirements.txt
# From Zoomcamp specifications
pandas==1.5.2
prefect==2.7.7
prefect-sqlalchemy==0.2.2
prefect-gcp[cloud_storage]==0.2.4
protobuf==4.21.11
pyarrow==10.0.1
pandas-gbq==0.18.1

# Others
requests
docker
psycopg2
bs4
lxml  # for bs4 Wikipedia HTML table reading

# try to force install
pydantic==1.10.0